fix other discord bots
