+++ bots/discord_basin_status_bot.py
@@ -3,8 +3,10 @@ import logging.handlers
 import os
 import signal
 
-from bots import util
+from constants.config import ONE_HUNDRED_MEGABYTES
 
+from bots import util
+from monitors.preview.basin import BasinStatusPreviewMonitor
 
 if __name__ == "__main__":
     logging.basicConfig(
@@ -31,7 +33,7 @@ if __name__ == "__main__":
         token = os.environ["DISCORD_BOT_TOKEN"]
         prod = False
 
-    client = util.DiscordSidebarClient(util.BasinStatusPreviewMonitor)
+    client = util.DiscordSidebarClient(BasinStatusPreviewMonitor)
 
     try:
         client.run(token)
