+++ bots/discord_eth_bot.py
@@ -5,8 +5,10 @@ import logging.handlers
 import os
 import signal
 
-from bots import util
+from constants.config import ONE_HUNDRED_MEGABYTES
 
+from bots import util
+from monitors.preview.eth import EthPreviewMonitor
 
 if __name__ == "__main__":
     logging.basicConfig(
@@ -33,7 +35,7 @@ if __name__ == "__main__":
         token = os.environ["DISCORD_BOT_TOKEN"]
         prod = False
 
-    client = util.DiscordSidebarClient(util.EthPreviewMonitor)
+    client = util.DiscordSidebarClient(EthPreviewMonitor)
 
     try:
         client.run(token)
