+++ bots/discord_nft_bot.py
@@ -5,8 +5,10 @@ import logging.handlers
 import os
 import signal
 
-from bots import util
+from constants.config import ONE_HUNDRED_MEGABYTES
 
+from bots import util
+from monitors.preview.nft import NFTPreviewMonitor
 
 if __name__ == "__main__":
     logging.basicConfig(
@@ -33,7 +35,7 @@ if __name__ == "__main__":
         token = os.environ["DISCORD_BOT_TOKEN"]
         prod = False
 
-    client = util.DiscordSidebarClient(util.NFTPreviewMonitor)
+    client = util.DiscordSidebarClient(NFTPreviewMonitor)
 
     try:
         client.run(token)
