+++ bots/discord_snapshot_bot.py
@@ -5,8 +5,10 @@ import logging.handlers
 import os
 import signal
 
-from bots import util
+from constants.config import ONE_HUNDRED_MEGABYTES
 
+from bots import util
+from monitors.preview.snapshot import SnapshotPreviewMonitor
 
 if __name__ == "__main__":
     logging.basicConfig(
@@ -33,7 +35,7 @@ if __name__ == "__main__":
         token = os.environ["DISCORD_BOT_TOKEN"]
         prod = False
 
-    client = util.DiscordSidebarClient(util.SnapshotPreviewMonitor)
+    client = util.DiscordSidebarClient(SnapshotPreviewMonitor)
 
     try:
         client.run(token)
